requirements.txt
requests==2.31.0
pandas==2.1.4
python-dotenv==1.0.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
schedule==1.2.0
tqdm==4.66.1